优化：学员组id由自增id转雪花id

+++ Sourcecode/Song.Entities/Accounts.cs
@@ -102,7 +102,7 @@ namespace Song.Entities {
     		
     		protected Int32 _Dep_Id;
     		
-    		protected Int32 _Sts_ID;
+    		protected Int64 _Sts_ID;
     		
     		protected String _Sts_Name;
     		
@@ -586,7 +586,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Sts_ID {
+    		public Int64 Sts_ID {
     			get {
     				return this._Sts_ID;
     			}
@@ -946,7 +946,7 @@ namespace Song.Entities {
     				this._Dep_Id = reader.GetInt32(_.Dep_Id);
     			}
     			if ((false == reader.IsDBNull(_.Sts_ID))) {
-    				this._Sts_ID = reader.GetInt32(_.Sts_ID);
+    				this._Sts_ID = reader.GetInt64(_.Sts_ID);
     			}
     			if ((false == reader.IsDBNull(_.Sts_Name))) {
     				this._Sts_Name = reader.GetString(_.Sts_Name);
@@ -1228,7 +1228,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Dep_Id = new WeiSha.Data.Field<Accounts>("Dep_Id");
     			
     			/// <summary>
-    			/// 字段名：Sts_ID - 数据类型：Int32
+    			/// 字段名：Sts_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Sts_ID = new WeiSha.Data.Field<Accounts>("Sts_ID");
