+++ Sourcecode/Song.Entities/Student.cs
@@ -84,7 +84,7 @@ namespace Song.Entities {
     		
     		protected String _Org_Name;
     		
-    		protected Int32 _Sts_ID;
+    		protected Int64 _Sts_ID;
     		
     		protected String _Sts_Name;
     		
@@ -484,7 +484,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Sts_ID {
+    		public Int64 Sts_ID {
     			get {
     				return this._Sts_ID;
     			}
@@ -835,7 +835,7 @@ namespace Song.Entities {
     				this._Org_Name = reader.GetString(_.Org_Name);
     			}
     			if ((false == reader.IsDBNull(_.Sts_ID))) {
-    				this._Sts_ID = reader.GetInt32(_.Sts_ID);
+    				this._Sts_ID = reader.GetInt64(_.Sts_ID);
     			}
     			if ((false == reader.IsDBNull(_.Sts_Name))) {
     				this._Sts_Name = reader.GetString(_.Sts_Name);
@@ -1081,7 +1081,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Org_Name = new WeiSha.Data.Field<Student>("Org_Name");
     			
     			/// <summary>
-    			/// 字段名：Sts_ID - 数据类型：Int32
+    			/// 字段名：Sts_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Sts_ID = new WeiSha.Data.Field<Student>("Sts_ID");
