+++ Sourcecode/Song.ServiceImpls/ExaminationCom.cs
@@ -825,7 +825,7 @@ namespace Song.ServiceImpls
         /// <param name="id">当前考试主题的ID</param>
         /// <param name="stsid">学生分组的id，为0时取所有，为-1时取不在组的学员，大于0则取当前组学员</param>
         /// <returns></returns>
-        public DataTable Result4Theme(int examid, int stsid)
+        public DataTable Result4Theme(int examid, long stsid)
         {
             Examination theme = this.ExamSingle(examid);
             Examination[] exams = this.ExamItem(theme.Exam_UID);    //当前考试下的多场考试
@@ -967,7 +967,7 @@ namespace Song.ServiceImpls
         /// <param name="stsid">学生分组的id，为0时取所有，为-1时取不在组的学员，大于0则取当前组学员</param>
         /// <param name="isAll">是否取所有人员（含缺考人员）,false为仅参考人员</param>
         /// <returns></returns>
-        public DataTable Result4Theme(int id, int stsid, bool isAll)
+        public DataTable Result4Theme(int id, long stsid, bool isAll)
         {
             if (!isAll) return Result4Theme(id, stsid);
             Examination theme = this.ExamSingle(id);
