+++ Sourcecode/Song.ServiceInterfaces/IExamination.cs
@@ -241,7 +241,7 @@ namespace Song.ServiceInterfaces
         /// <param name="id">当前考试主题的ID</param>
         /// <param name="stsid">学生分组的id，为0时取所有，为-1时取不在组的学员，大于0则取当前组学员</param>
         /// <returns></returns>
-        DataTable Result4Theme(int examid, int stsid);
+        DataTable Result4Theme(int examid, long stsid);
         /// <summary>
         /// 考试主题下的所有参考人员成绩
         /// </summary>
@@ -256,7 +256,7 @@ namespace Song.ServiceInterfaces
         /// <param name="stsid">学生分组的id，为0时取所有，为-1时取不在组的学员，大于0则取当前组学员</param>
         /// <param name="isAll">是否取所有人员（含缺考人员）,false为仅参考人员</param>
         /// <returns></returns>
-        DataTable Result4Theme(int id, int stsid, bool isAll);
+        DataTable Result4Theme(int id, long stsid, bool isAll);
         /// <summary>
         /// 当前考试主题下的各学员分组成绩排行
         /// </summary>
